+++ changelog.md
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+dev
+------------------
+- Predicate.content attribute no longer includes an extra placeholder for a "quantity"
+
 0.4.1 (2021-01-02)
 ------------------
 - Increase minimum Legislice version to 0.4.1